[MOVE] DateAndTime 코드 분리 #25
